requirements.txt
Django~=4.2.2
Djangorestframework~=3.14.0
pytz~=2023.3
